[exp] use uint64_t for mask to support more VCIs This patch changes the integer type of VCI index mask to support more VCIs (32 -> 64).

+++ src/include/mpir_thread.h
@@ -122,7 +122,7 @@ extern MPIU_exp_data_t g_MPIU_exp_data;
 
 typedef struct {
     char dummy1[64];
-    int vci_mask;
+    uint64_t vci_mask;
 #if defined(VCIEXP_LOCK_PTHREADS)
     int local_tid;
 #endif
